Fix lock-up if stopping playback in the middle of a file read
First we disable audio thread's access to shared data by setting its
PlayState to Restart (from the audio thread). Then in the next async
thread we reset the buffers.
Otherwise, if we reset prebuff or frames_to_read from the audio thread,
it could be interrupting the async thread in teh middle of a read, which
will have it write to the prebuff in the wrong place